WebService/requirements.txt
fastapi
uvicorn
pydantic
grpcio
grpcio-tools
protobuf